プロジェクトを分割
