+++ CpLibrary.sln
@@ -3,7 +3,9 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.30517.126
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CpLibrary", "CpLibrary\CpLibrary.csproj", "{1B906B87-5B74-4FA7-8FEC-56501C82AB41}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CpLibrary", "CpLibrary\CpLibrary.csproj", "{1B906B87-5B74-4FA7-8FEC-56501C82AB41}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "CpProject", "CpProject\CpProject.csproj", "{8F381C27-A19D-41F7-9DB6-A5882393A375}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{1B906B87-5B74-4FA7-8FEC-56501C82AB41}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{1B906B87-5B74-4FA7-8FEC-56501C82AB41}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{1B906B87-5B74-4FA7-8FEC-56501C82AB41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8F381C27-A19D-41F7-9DB6-A5882393A375}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8F381C27-A19D-41F7-9DB6-A5882393A375}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8F381C27-A19D-41F7-9DB6-A5882393A375}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8F381C27-A19D-41F7-9DB6-A5882393A375}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
