+++ CpLibrary/CpLibrary.csproj
@@ -1,21 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<OutputType>Exe</OutputType>
+		<OutputType>Library</OutputType>
 		<TargetFramework>net5.0</TargetFramework>
 		<LangVersion>8</LangVersion>
 		<AllowUnsafeBlocks>true</AllowUnsafeBlocks>
 		<NoWarn>1701;1702;IDE0038;IDE0044;CS0649;CA1034;CA1031;CA1815;CA1822;CA1819;IDE0051;IDE0052;IDE0059;IDE1006;IDE0019;IDE0017;IDE0034;IDE0056</NoWarn>
-		<StartupObject>CpLibrary.Program</StartupObject>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|AnyCPU'">
-		<OutputPath></OutputPath>
-	</PropertyGroup>
-
-	<PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
-		<OutputPath></OutputPath>
-		<CheckForOverflowUnderflow>true</CheckForOverflowUnderflow>
 	</PropertyGroup>
 
 	<ItemGroup>
@@ -26,7 +16,6 @@
 		</PackageReference>
 		<PackageReference Include="CommandLineParser" Version="2.9.0-preview1" />
 		<PackageReference Include="Microsoft.CodeAnalysis" Version="3.8.0" />
-		<PackageReference Include="SourceExpander" Version="2.5.0" />
 		<PackageReference Include="SourceExpander.Embedder" Version="2.5.0">
 		  <PrivateAssets>all</PrivateAssets>
 		  <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
