--- CpLibrary/Workspace/Info.json
@@ -1,4 +1,0 @@
-﻿{
-  "contestSite": "atcoder",
-  "contestName": "abc187"
-}
