+++ CpProject/Contest/B.cs
@@ -1,0 +1,33 @@
+﻿using System;
+using System.IO;
+using System.Collections.Generic;
+using System.Linq;
+using System.Numerics;
+using System.Runtime.CompilerServices;
+using System.Text;
+using System.Globalization;
+using System.Threading;
+using System.Diagnostics;
+
+using static System.Math;
+
+namespace CpLibrary.Contest
+{
+	public class B
+	{
+		public B()
+		{
+
+		}
+
+		public void Run()
+		{
+
+		}
+
+		public void Solve()
+		{
+
+		}
+	}
+}
