+++ CpProject/Workspace/Info.json
@@ -1,0 +1,4 @@
+﻿{
+  "contestSite": "atcoder",
+  "contestName": "abc187"
+}
